refactor: polish tools package
