+++ api/chatbot/tools/weather/__init__.py
@@ -1,0 +1,3 @@
+from .openmeteo import WeatherTool
+
+__all__ = ["WeatherTool"]
